+++ src/lib.rs
@@ -1,7 +1,3 @@
-pub mod dom_struct;
-pub mod dom_text;
-pub mod dom_vec;
+pub mod editor;
 pub mod utils;
 pub mod web_support;
-
-pub mod editor;
