+++ src/main.rs
@@ -1,5 +1,6 @@
 use anyhow::Result;
-use codillon::{dom_struct::DomStruct, editor::Editor, web_support::DocumentHandle};
+use codillon::editor::Editor;
+use codillon::web_support::{DocumentHandle, components::DomStruct};
 use std::cell::RefCell;
 
 type Body = DomStruct<(Editor, ()), web_sys::HtmlBodyElement>;
