+++ src/web_support/components.rs
@@ -1,0 +1,7 @@
+mod dom_struct;
+mod dom_text;
+mod dom_vec;
+
+pub use dom_struct::*;
+pub use dom_text::*;
+pub use dom_vec::*;
